requirements/requirements.txt
scipy
scikit-learn
ruamel.yaml
multiprocess
setproctitle
joblib
colorlog

autolab-core
autolab-perception
visualization
opencv-python==4.2.0.32
opencv-contrib-python==4.2.0.32
scikit-image<=0.14.2
gputil
tensorflow==1.13.1

pylibfreenect2
tensorflow-estimator >= 1.13.0, <1.14.0rc0
pyquaternion

keras

ipython==5.5.0
scikit-video
ffmpeg-python

Pillow
docutils
pyserial>=3.4
cython
psutil
matplotlib<=2.2.0
numpy
cycler